fix: bootstrap error

+++ scripts/bootstrap.ts
@@ -55,12 +55,12 @@ import { $, argv, chalk, fs, path } from 'zx';
 
       // package.json
       const pkgPkgJSONPath = path.join(pkgDir, 'package.json');
-      const pkgPkgJSON = fs.existsSync(pkgPkgJSONPath)
+      const hasPkgJSON = fs.existsSync(pkgPkgJSONPath);
+      const pkgPkgJSON = hasPkgJSON
         ? require(pkgPkgJSONPath)
         : {};
       fs.writeJSONSync(
         pkgPkgJSONPath,
-
         Object.assign(
           {
             name,
@@ -89,25 +89,29 @@ import { $, argv, chalk, fs, path } from 'zx';
             },
           },
           {
-            authors: pkgPkgJSON.authors,
-            bin: pkgPkgJSON.bin,
-            files: pkgPkgJSON.files,
-            scripts: pkgPkgJSON.scripts,
-            description: pkgPkgJSON.description,
-            dependencies: pkgPkgJSON.dependencies,
-            devDependencies: pkgPkgJSON.devDependencies,
-            compiledConfig: pkgPkgJSON.compiledConfig,
+            ...(hasPkgJSON ? {
+              authors: pkgPkgJSON.authors,
+              bin: pkgPkgJSON.bin,
+              files: pkgPkgJSON.files,
+              scripts: pkgPkgJSON.scripts,
+              description: pkgPkgJSON.description,
+              dependencies: pkgPkgJSON.dependencies,
+              devDependencies: pkgPkgJSON.devDependencies,
+              compiledConfig: pkgPkgJSON.compiledConfig,
+            } : {})
           },
         ),
         { spaces: '  ' },
       );
 
       // README.md
-      await fs.writeFile(
-        path.join(pkgDir, 'README.md'),
-        `# ${name}\n`,
-        'utf-8',
-      );
+      if (!fs.existsSync(path.join(pkgDir, 'README.md'))) {
+        await fs.writeFile(
+          path.join(pkgDir, 'README.md'),
+          `# ${name}\n`,
+          'utf-8',
+        );
+      }
 
       // tsconfig.json
       await fs.writeFile(
